re-organize code a bit
